# openapi.yaml
openapi: 3.1.0
info:
  title: wa-hub-demo — WhatsApp HTTP API
  version: 0.2.0
  description: >
    Self-hosted WhatsApp HTTP API (Baileys). REST + JSON, with HMAC-signed outbound webhooks.
    PRECONDITION: a WhatsApp device must be paired once (scan a QR at GET /pair) before any send
    works — otherwise send routes return 503 not_connected. Confirm readiness with
    GET /api/instance/status (connection must be "connected"). See docs/INTEGRATION.md.
  license:
    name: MIT
servers:
  - url: https://wa.yourdomain.com
    description: Your stable subdomain (recommended) — or the ephemeral https://<random>.trycloudflare.com tunnel URL
  - url: http://127.0.0.1:3060
    description: On the server itself (loopback)
security:
  - bearerAuth: []
...
paths:
...
  /api/groups:
    get:
      tags: [groups]
      summary: List groups the paired account participates in
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  groups:
                    type: array
                    items:
                      type: object
                      properties:
                        jid: { type: string }
                        name: { type: string }
                        participants: { type: integer }
                        owner: { type: string, nullable: true }
                        creation: { type: integer, nullable: true }
                        announce: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
    post:
      tags: [groups]
      summary: Create a group
      description: >
        ⚠ Bulk-creating groups or adding strangers is a top ban signal — use only with consenting participants.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject]
              properties:
                subject: { type: string, minLength: 1, maxLength: 100 }
                participants: { type: array, maxItems: 50, default: [], items: { type: string } }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  jid: { type: string }
                  subject: { type: string }
                  participants: { type: integer }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/groups/{jid}:
    get:
      tags: [groups]
      summary: Full group metadata
      parameters:
        - { name: jid, in: path, required: true, schema: { type: string }, description: "15+ digit group id or <id>@g.us" }
      responses:
        '200': { description: Group metadata, content: { application/json: { schema: { type: object } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/groups/{jid}/participants:
    post:
      tags: [groups]
      summary: Add / remove / promote / demote participants
      parameters:
        - { name: jid, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                add: { type: array, maxItems: 50, items: { type: string } }
                remove: { type: array, maxItems: 50, items: { type: string } }
                promote: { type: array, maxItems: 50, items: { type: string } }
                demote: { type: array, maxItems: 50, items: { type: string } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/groups/{jid}/leave:
    post:
      tags: [groups]
      summary: Leave a group
      parameters:
        - { name: jid, in: path, required: true, schema: { type: string }, description: "15+ digit group id or <id>@g.us" }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/groups/{jid}/subject:
    put:
      tags: [groups]
      summary: Rename a group (set its subject)
      parameters:
        - { name: jid, in: path, required: true, schema: { type: string }, description: "15+ digit group id or <id>@g.us" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject]
              properties:
                subject: { type: string, minLength: 1, maxLength: 100 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/groups/{jid}/description:
    put:
      tags: [groups]
      summary: Set the group description (empty/omitted clears it)
      parameters:
        - { name: jid, in: path, required: true, schema: { type: string }, description: "15+ digit group id or <id>@g.us" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description: { type: string, maxLength: 2048 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/groups/{jid}/settings:
    put:
      tags: [groups]
      summary: Lock/announce settings (who can post / edit group info)
      parameters:
        - { name: jid, in: path, required: true, schema: { type: string }, description: "15+ digit group id or <id>@g.us" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [setting]
              properties:
                setting: { type: string, enum: [announcement, not_announcement, locked, unlocked] }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/groups/{jid}/ephemeral:
    put:
      tags: [groups]
      summary: Toggle disappearing messages
      description: "seconds: 0 off, 86400 (24h), 604800 (7d), 7776000 (90d)."
      parameters:
        - { name: jid, in: path, required: true, schema: { type: string }, description: "15+ digit group id or <id>@g.us" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [seconds]
              properties:
                seconds: { type: integer, minimum: 0 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/groups/{jid}/member-add-mode:
    put:
      tags: [groups]
      summary: Who may add members (admins only vs all members)
      parameters:
        - { name: jid, in: path, required: true, schema: { type: string }, description: "15+ digit group id or <id>@g.us" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mode]
              properties:
                mode: { type: string, enum: [admin_add, all_member_add] }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/groups/{jid}/join-approval:
    put:
      tags: [groups]
      summary: Require admin approval for joins via the invite link
      parameters:
        - { name: jid, in: path, required: true, schema: { type: string }, description: "15+ digit group id or <id>@g.us" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mode]
              properties:
                mode: { type: string, enum: [on, off] }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/groups/{jid}/invite:
    get:
      tags: [groups]
      summary: Get the group invite link/code (admin only)
      parameters:
        - { name: jid, in: path, required: true, schema: { type: string }, description: "15+ digit group id or <id>@g.us" }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  code: { type: string, nullable: true }
                  url: { type: string, nullable: true }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/groups/{jid}/invite/revoke:
    post:
      tags: [groups]
      summary: Revoke the current invite link and return the new one (admin only)
      parameters:
        - { name: jid, in: path, required: true, schema: { type: string }, description: "15+ digit group id or <id>@g.us" }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  code: { type: string, nullable: true }
                  url: { type: string, nullable: true }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/groups/{jid}/requests:
    get:
      tags: [groups]
      summary: List pending join requests (admin only)
      parameters:
        - { name: jid, in: path, required: true, schema: { type: string }, description: "15+ digit group id or <id>@g.us" }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  requests: { type: array, items: { type: object } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
    post:
      tags: [groups]
      summary: Approve or reject pending join requests (admin only)
      parameters:
        - { name: jid, in: path, required: true, schema: { type: string }, description: "15+ digit group id or <id>@g.us" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [participants, action]
              properties:
                participants: { type: array, minItems: 1, maxItems: 50, items: { type: string } }
                action: { type: string, enum: [approve, reject] }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean }, results: { type: array, items: { type: object } } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/groups/invite/{code}:
    get:
      tags: [groups]
      summary: Preview a group from an invite code (no join)
      parameters:
        - { name: code, in: path, required: true, schema: { type: string }, description: "The part after chat.whatsapp.com/ (or the full link)" }
      responses:
        '200': { description: Group metadata, content: { application/json: { schema: { type: object } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
  /api/groups/join:
    post:
      tags: [groups]
      summary: Join a group via invite code/link
      description: "⚠ Mass-joining is a spam signal — use sparingly."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, minLength: 6, description: "Invite code or full chat.whatsapp.com/ link" }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { ok: { type: boolean }, jid: { type: string } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/NotConnected' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: The HUB_TOKEN.
    adminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Required on PUT /webhook and POST /logout only when ADMIN_TOKEN is configured.
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
  responses:
...
    Unauthorized:
      description: Missing or invalid HUB_TOKEN
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    BadRequest:
      description: >
        Invalid body / recipient / JSON — e.g. {"error":"invalid_body"} (with Zod issues) or
        {"error":"invalid_json"}. Oversized base64 media (…Base64 over the 20 MB decoded cap)
        also lands here as {"error":"invalid_body"} — "file_too_large" appears only inside
        issues[].message, NOT as the top-level error.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    NotConnected:
      description: WhatsApp is not paired/connected yet
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }